Real-time multichannel mixer support and example
Can play WAV over WAV or MP3s, will automatically resample to the output
sample rate as chosen by the user.  Example added to play WAV over AAC.

Because it is driven off of the output IRQ it is not necessary for all
inputs to be generating at the same rate, or even at all.