Update Webradio URL to MP3 station
The web radio station used in the example has migrated from MP3 to
"MPEG ADTS, AAC, v4 LC, 22.05 kHz, monaural" stream which is not
compatible with the MP3 decoder.

Update to just point to the BBC for now.  Other sites from the
web radio collection listed in the comment still work.