Make ESP32I2S buffers <1023 silently
The IDF will truncate buffer lengths to 1023 words or less, but if it does
that the logic behind the availableForWrite tracking is no longer valid.

Divide the buffer length in half and double the buffer count until we do
fit within the legal range.  The upper layer codec will get add'l IRQ
callbacks, but it should be able to use the availableForWrite() to early
terminate when there isn't a full requested buffer available.