Allow for small delay on ESP32I2S write
Under experimentation it was seen that even after > 4K worth of space
was returned to the dma free pool (counting the i2s_event_data_t.size)
that a write of *less than that amount* could actually fail with a 0
timeout.

It's unclear to me how a write of less data than we have space for
could ever timeout.  I verified we got 5 IRQ callbacks of 512 32b words
each, meaning 10240 bytes of free space was available.  A write of
even 100 bytes failed with timeout==0 in this case.

Possibly the IDF has some lock taken (another core? another task that
got swapped?) when we're trying to write and a timeout would be needed
to task switch and clean things up.

In any case, adding a 100ms max delay on the i2s_data_write seems to
have cleared up the PlayAACFromROM and Webradio examples.

Also add an assertion in the AAC decode that the write to the
output renderer succeeded.