ESP32I2S write should try multiple DMA buffers, obey size constraints, allow for small timeout (#72)
* ESP32I2S write should try multiple DMA buffers

ESP32 i2s_channel_write seems to return after filling the current DMA
buffer, even if free add'l buffers could be written.  Move the retry
from MP3 decode to the ESP32I2S internal so all codecs can benefit.

See #65 for discussion

* Make ESP32I2S buffers <1023 silently

The IDF will truncate buffer lengths to 1023 words or less, but if it does
that the logic behind the availableForWrite tracking is no longer valid.

Divide the buffer length in half and double the buffer count until we do
fit within the legal range.  The upper layer codec will get add'l IRQ
callbacks, but it should be able to use the availableForWrite() to early
terminate when there isn't a full requested buffer available.

* Allow for small delay on ESP32I2S write

Under experimentation it was seen that even after > 4K worth of space
was returned to the dma free pool (counting the i2s_event_data_t.size)
that a write of *less than that amount* could actually fail with a 0
timeout.

It's unclear to me how a write of less data than we have space for
could ever timeout.  I verified we got 5 IRQ callbacks of 512 32b words
each, meaning 10240 bytes of free space was available.  A write of
even 100 bytes failed with timeout==0 in this case.

Possibly the IDF has some lock taken (another core? another task that
got swapped?) when we're trying to write and a timeout would be needed
to task switch and clean things up.

In any case, adding a 100ms max delay on the i2s_data_write seems to
have cleared up the PlayAACFromROM and Webradio examples.

Also add an assertion in the MP3/AAC decode that the write to the
output renderer succeeded.