Use lower-prio IRQ on Pico to decompress AAC/MP3
Fixes #84

On the Pico, when decoding AAC or MP3, it can spend a long time in
an IRQ state, blocking other IRQs from processing.  While we need
to run asynchronously, we don't need to block other IRQs.

When a DMA IRQ fires and calls us to generate the next frame,
fire another low priority IRQ and return immediately.  In the worst
case we still run ahead of the app with just a little overhead.
In the best case we don't block GPIO or USB interrupts.